Squashed commit of the following:
commit dc5800f4cdab4b7068b728ce15cae1ff1d5dd6dc
Merge: 69a6497 9968c95
Author: AlMkin <138128998+AlMkin@users.noreply.github.com>
Date:   Tue Mar 26 19:22:57 2024 +0300

    Merge pull request #11 from Trudovyereservy/feature/eslint-import-sort

    Feature/eslint import sort

commit 9968c95a085ea6e5ae97179ad84a5da988322b81
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Mon Mar 25 19:48:57 2024 +0700

    feat: добавил правило в линтер ограничивающее количество пустых строк до максимум одной

commit 7efd298f956bb283ccdcff493597ab910156fb34
Merge: ab25abe 69a6497
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Mon Mar 25 19:08:50 2024 +0700

    Merge branch 'develop' of github.com:Trudovyereservy/frontend_part into feature/eslint-import-sort

commit ab25abef08e032e791c443f5ed56a976f6dc8a41
Merge: 92f993d 62bce03
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Sat Mar 23 14:18:59 2024 +0700

    Merge branch 'feature/eslint-import-sort' of github.com:Trudovyereservy/frontend_part into feature/eslint-import-sort

commit 92f993d65e05b87cad99d6bfc4d30440b0ab2162
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Sat Mar 23 14:13:33 2024 +0700

    feat: добавил правило пустой строки после директив и исправил везде yarn lint --fix

commit 62bce034bcde734c48eb32226d74219434fafd51
Merge: aa8b781 359e06a
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Sat Mar 23 13:16:21 2024 +0700

    Merge branch 'feature/eslint-import-sort' of github.com:Trudovyereservy/frontend_part into feature/eslint-import-sort

commit aa8b781f036187f45c85dc674460784ee3eab89a
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Sat Mar 23 13:16:09 2024 +0700

    fix: поправил название constant в импорте, добавил правило пустой строки после директив

commit 69a6497d07b4793560cb4b191aa99bbe9a2310da
Author: Pavel Demin <pavlikdemin@yandex.ru>
Date:   Sat Mar 23 01:35:51 2024 +0300

    Поправил имя constants, поправил по мелочи файлы

commit 359e06a0baf2441390a1627095fdb66f7fd434e4
Merge: 6ba9071 14b64e3
Author: Glushenkov Konstantin <glukosalex@gmail.com>
Date:   Thu Mar 21 22:05:02 2024 +0700

    Merge branch 'develop' into feature/eslint-import-sort

commit 14b64e3888846e4ca67ea3c044f2db2a9003d592
Author: Pavel Demin <pavlikdemin@yandex.ru>
Date:   Thu Mar 21 00:09:24 2024 +0300

    Сделал один файл Constans, удалил в форме картинки по 7 МБ, скорректировал компонент Input, в константах поправил уникальность id

commit 6ba90717a5982463bd786474dce1c228b066de5d
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Wed Mar 20 23:07:03 2024 +0700

    fix: fix path to images

commit e9c3a247af466cf70f1b6aaa6d0e920a1cfb2343
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Wed Mar 20 23:02:57 2024 +0700

    style: fix parent paths to alias '@'

commit d70a6801eefa56f4386efacc6f952c08d997ed0b
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Wed Mar 20 23:02:04 2024 +0700

    style: add eslint rule import/newline-after-import

commit fba2499ac5f1e5d85bce8af328f4d8ad8e8ca0b7
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Wed Mar 20 22:33:12 2024 +0700

    style: auto sort all imports

commit c56c179b63378a53ba1fce00b7b4064aa7343af9
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Wed Mar 20 22:32:31 2024 +0700

    style: config eslint and prettier

commit f969aecdaff18e5ac8ad0b770885a0bfc63579c9
Author: GluKosAlex <glukosalex@gmail.com>
Date:   Wed Mar 20 22:31:08 2024 +0700

    build: install simple-import-sort plugin
