+++ src/app/coachs/page.tsx
@@ -1,7 +1,8 @@
 'use client';
-import { CoachList } from '../../components/CoachCards/CoachList/CoachList';
-import { coachCards } from '@/utils/constans';
+
+import { CoachList } from '@/components/CoachCards/CoachList/CoachList';
 import { DescriptionPages } from '@/components/DescriptionPages/DescriptionPages';
+import { coachCards, descriptionPages } from '@/utils/constants';
 
 export default function CoachsPage() {
   return (
@@ -10,7 +11,7 @@ export default function CoachsPage() {
         <title>Трудовые резервы | Тренерский состав</title>
         <meta name="title" content="Блог" />
       </head>
-      <DescriptionPages />
+      <DescriptionPages descriptionPages={descriptionPages} />
       <CoachList coachCards={coachCards} />
     </>
   );
