+++ src/app/contacts/page.tsx
@@ -1,5 +1,7 @@
 'use client';
+
 import { DescriptionPages } from '@/components/DescriptionPages/DescriptionPages';
+import { descriptionPages } from "@/utils/constants";
 
 export default function ContactsPage() {
   return (
@@ -8,7 +10,7 @@ export default function ContactsPage() {
         <title>Трудовые резервы | Контакты</title>
         <meta name="title" content="Блог" />
       </head>
-      <DescriptionPages />
+      <DescriptionPages descriptionPages={descriptionPages} />
     </>
   );
 }
