+++ src/app/news/page.tsx
@@ -1,5 +1,7 @@
 'use client';
+
 import { DescriptionPages } from "@/components/DescriptionPages/DescriptionPages";
+import { descriptionPages } from "@/utils/constants";
 
 export default function NewsPage() {
   return (
@@ -8,7 +10,7 @@ export default function NewsPage() {
         <title>Трудовые резервы | Новости</title>
         <meta name="title" content="Блог" />
       </head>
-      <DescriptionPages />
+      <DescriptionPages descriptionPages={descriptionPages} />
     </>
   );
 }
