+++ src/app/page.tsx
@@ -1,10 +1,13 @@
 'use client';
-import styles from './page.module.scss';
+
 import { useState } from 'react';
-import { MainBanner } from '@/components/main-banner/main-banner';
+
+import { CardsList } from '@/components/CardsList/CardsList';
 import { Faq } from "@/components/Faq/Faq";
 import  Form  from "@/components/Form/Form";
-import { CardsList } from '../components/CardsList/CardsList';
+import { MainBanner } from '@/components/main-banner/main-banner';
+
+import styles from './page.module.scss';
 
 export default function Home() {
   const [closeMenu, setCloseMenu] = useState(false);
