+++ src/components/BlockCards/Card/Card.tsx
@@ -1,6 +1,8 @@
-import Link from 'next/link';
 import Image from 'next/image';
+import Link from 'next/link';
+
 import { ICardBlogProps } from './Card.props';
+
 import styles from './Card.module.scss';
 
 const Card = ({ title, shortDescription, imgUrl, linkUrl }: ICardBlogProps) => {
