+++ src/components/CardsList/CardsList.tsx
@@ -1,6 +1,8 @@
+import { Card } from '@/components/Card/Card';
+import { testCards } from '@/utils/constants';
+
 import styles from './CardsList.module.scss';
-import { Card } from '../Card/Card';
-import { testCards } from '@/utils/constatns';
+
 const CardsList = () => {
   return (
     <section className={styles.cardslist}>
