+++ src/components/CoachCards/Card/Card.tsx
@@ -1,8 +1,11 @@
-import Image from 'next/image';
-import { ICardCoachProps } from './Card.props';
-import styles from './Card.module.scss';
 import { useState } from 'react';
+
 import classNames from 'classnames';
+import Image from 'next/image';
+
+import { ICardCoachProps } from './Card.props';
+
+import styles from './Card.module.scss';
 
 const Card = ({ photo, name, surname, directions, achievements, patronymic }: ICardCoachProps) => {
   const [openInfo, setOpenInfo] = useState(true);
