+++ src/components/CoachCards/CoachList/CoachList.tsx
@@ -1,9 +1,11 @@
 import {useMemo} from 'react';
+
+import { Card } from '@/components/CoachCards/Card/Card';
+import { ICardCoachProps } from '@/components/CoachCards/Card/Card.props';
+import { useCardCount } from '@/hooks/useCardCount';
+import useWindowSize from '@/hooks/useWindowSize';
+
 import styles from './CoachList.module.scss';
-import { Card } from '../Card/Card';
-import { ICardCoachProps } from '../Card/Card.props';
-import useWindowSize from '../../../hooks/useWindowSize';
-import { useCardCount } from '../../../hooks/useCardCount';
 
 const CoachList = ({ coachCards }: { coachCards: ICardCoachProps[] }) => {
   const width: number = useWindowSize();
