+++ src/components/Faq/Faq.tsx
@@ -1,6 +1,7 @@
+import { FaqItem } from '@/components/FaqItem/FaqItem';
+import { testFaqItems } from '@/utils/constants';
+
 import styles from './Faq.module.scss';
-import { FaqItem } from '../FaqItem/FaqItem';
-import { testFaqItems } from '../../utils/constants';
 
 const Faq = () => {
   return (
