+++ src/components/Footer/Footer.tsx
@@ -1,12 +1,15 @@
 'use client';
-import styles from './footer.module.scss';
-import Link from 'next/link';
-import { Button } from '../Button/Button';
-import { linksItems, buttonsNames } from '../../utils/constans';
-import { Input } from '../Input/Input';
-import { useForm } from 'react-hook-form';
-import { LinksBlock } from '../Links/LinksBlock';
+
 import classNames from 'classnames';
+import Link from 'next/link';
+import { useForm } from 'react-hook-form';
+
+import { Button } from '@/components/Button/Button';
+import { Input } from '@/components/Input/Input';
+import { LinksBlock } from '@/components/Links/LinksBlock';
+import { linksItems, buttonsNames } from '@/utils/constants';
+
+import styles from './footer.module.scss';
 
 const Footer = () => {
   const {
