+++ src/components/Header/Header.tsx
@@ -1,10 +1,15 @@
 'use client';
-import Link from 'next/link';
-import styles from './header.module.scss';
-import { headerLinks } from '@/utils/constans';
-import { LinksHeader } from './LinksHeader/LinksHeader';
-import Menu from '@/components/Menu/Menu';
+
 import { useState } from 'react';
+
+import Link from 'next/link';
+
+import Menu from '@/components/Menu/Menu';
+import { headerLinks } from '@/utils/constants';
+
+import { LinksHeader } from './LinksHeader/LinksHeader';
+
+import styles from './header.module.scss';
 
 const Header = () => {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
