+++ src/components/Input/Input.tsx
@@ -1,15 +1,21 @@
 import { InputProps } from './Input.props';
 
-const Input = (props: InputProps) => {
-  const { className, register, nameInput } = props;
+const Input = ({ className, register, nameInput, pattern, ...inputProps }: InputProps) => {
+  const { rest, ...restProps } = inputProps;
 
   return (
     <input
-    type="text"
-    placeholder="Phone"
-    className={className}
-    {...register(nameInput, { required: 'Введите телефон в формате +7-***-**-**', pattern: /^(\s*)?(\+)?([- _():=+]?\d[- _():=+]?){10,14}(\s*)?$/i })}
-    {...props}
+      type="text"
+      placeholder="Phone"
+      className={className}
+      {...register(nameInput, {
+        required: 'Введите телефон в формате +7-***-**-**',
+        pattern: {
+          value: /^(\s*)?(\+)?([- _():=+]?\d[- _():=+]?){10,14}(\s*)?$/,
+          message: 'Неверный формат телефона',
+        },
+      })}
+      {...restProps}
     />
   );
 };
