+++ src/components/Menu/Menu.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
-import styles from './menu.module.scss';
+
+import classNames from 'classnames';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import { navLinksMenu } from '@/utils/constans';
-import classNames from 'classnames';
+
+import { navLinksMenu } from '@/utils/constants';
+
+import styles from './menu.module.scss';
 
 const Menu = ({ handler, handleToggleMenu }) => {
   const closeMenu = () => {
