+++ src/components/main-banner/main-banner.tsx
@@ -1,10 +1,13 @@
 'use client';
-import styles from '../main-banner/main-banner.module.scss';
-import picturelogo from '../../../public/Mask group.png';
-import picturelogo_mobile from '../../../public/Mask group_394px.png';
+
 import Image from 'next/image';
-import { Button } from '../Button/Button';
-import { buttonsNames } from '@/utils/constans';
+
+import picturelogo from '@/../public/Mask group.png';
+import picturelogo_mobile from '@/../public/Mask group_394px.png';
+import { Button } from '@/components/Button/Button';
+import { buttonsNames } from '@/utils/constants';
+
+import styles from '../main-banner/main-banner.module.scss';
 
 const MainBanner = () => {
   return (
