+++ src/hooks/useCardCount.ts
@@ -1,4 +1,5 @@
 import { useMemo } from 'react';
+
 import {
     BREAKPOINT_DESKTOP,
     BREAKPOINT_LAPTOP,
